+++ combustion_ble/__init__.py
@@ -2,12 +2,18 @@
 
 """Top-level package for cobustion_ble."""
 
+from combustion_ble.ble_data.advertising_data import (
+    AdvertisingData,
+    CombustionProductType,
+)
+from combustion_ble.ble_data.mode_id import ProbeColor, ProbeID, ProbeMode
 from combustion_ble.ble_data.prediction_state import PredictionState
 from combustion_ble.ble_data.prediction_status import (
     PredictionMode,
     PredictionStatus,
     PredictionType,
 )
+from combustion_ble.ble_manager import BluetoothMode
 from combustion_ble.device_manager import DeviceManager
 from combustion_ble.devices.probe import VirtualTemperatures
 from combustion_ble.version import VERSION, VERSION_SHORT
@@ -15,11 +21,17 @@ from combustion_ble.version import VERSION, VERSION_SHORT
 __all__ = [
     "VERSION",
     "VERSION_SHORT",
+    "AdvertisingData",
+    "BluetoothMode",
+    "CombustionProductType",
     "DeviceManager",
     "devices",
     "PredictionMode",
     "PredictionStatus",
     "PredictionType",
     "PredictionState",
+    "ProbeID",
+    "ProbeColor",
+    "ProbeMode",
     "VirtualTemperatures",
 ]