+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- Allow externally managed `BleakScanner`s to be used via new `BluetoothMode` parameter. (#20)
 
 ## [v0.2.1](https://github.com/legrego/combustion_ble/releases/tag/v0.2.1) - 2024-01-22
 - Allow BleManager to be stopped and restarted (#14)